requirements.txt
Flask>=3.0
gunicorn>=21.2
python-dotenv>=1.0
openai>=1.0.0